[ADD] 查询用户信息

+++ base/io/response.go
@@ -74,8 +74,8 @@ type UserInfoResp struct {
 	Response
 	ID            int64  `json:"id"`
 	Name          string `json:"name"`
-	FollowCount   int    `json:"follow_count"`
-	FollowerCount int    `json:"follower_count"`
+	FollowCount   int64  `json:"follow_count"`
+	FollowerCount int64  `json:"follower_count"`
 	IsFollow      bool   `json:"is_follow"`
 }
